Shared class for 'target instant or sorcery card in your graveyard gains flashback until end of turn. The flashback cost is equal to its mana cost'
