+++ Mage.Sets/src/mage/cards/f/Flashback.java
@@ -2,19 +2,11 @@ package mage.cards.f;
 
 import java.util.UUID;
 
-import mage.abilities.Ability;
-import mage.abilities.effects.ContinuousEffectImpl;
-import mage.abilities.keyword.FlashbackAbility;
-import mage.cards.Card;
+import mage.abilities.effects.common.continuous.GainFlashbackTargetEffect;
 import mage.cards.CardImpl;
 import mage.cards.CardSetInfo;
 import mage.constants.CardType;
-import mage.constants.Duration;
-import mage.constants.Layer;
-import mage.constants.Outcome;
-import mage.constants.SubLayer;
 import mage.filter.StaticFilters;
-import mage.game.Game;
 import mage.target.common.TargetCardInYourGraveyard;
 
 /**
@@ -27,7 +19,7 @@ public final class Flashback extends CardImpl {
         super(ownerId, setInfo, new CardType[]{CardType.INSTANT}, "{R}");
 
         // Target instant or sorcery card in your graveyard gains flashback until end of turn. The flashback cost is equal to its mana cost.
-        this.getSpellAbility().addEffect(new FlashbackEffect());
+        this.getSpellAbility().addEffect(new GainFlashbackTargetEffect());
         this.getSpellAbility().addTarget(new TargetCardInYourGraveyard(StaticFilters.FILTER_CARD_INSTANT_OR_SORCERY));
     }
 
@@ -38,35 +30,5 @@ public final class Flashback extends CardImpl {
     @Override
     public Flashback copy() {
         return new Flashback(this);
-    }
-}
-
-class FlashbackEffect extends ContinuousEffectImpl {
-
-    FlashbackEffect() {
-        super(Duration.EndOfTurn, Layer.AbilityAddingRemovingEffects_6, SubLayer.NA, Outcome.AddAbility);
-        this.staticText = "target instant or sorcery card in your graveyard gains flashback until end of turn. The flashback cost is equal to its mana cost";
-    }
-
-    private FlashbackEffect(final FlashbackEffect effect) {
-        super(effect);
-    }
-
-    @Override
-    public FlashbackEffect copy() {
-        return new FlashbackEffect(this);
-    }
-
-    @Override
-    public boolean apply(Game game, Ability source) {
-        Card card = game.getCard(getTargetPointer().getFirst(game, source));
-        if (card != null) {
-            FlashbackAbility ability = new FlashbackAbility(card, card.getManaCost());
-            ability.setSourceId(card.getId());
-            ability.setControllerId(card.getOwnerId());
-            game.getState().addOtherAbility(card, ability);
-            return true;
-        }
-        return false;
     }
 }
